@@ -11,32 +11,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "18.2.13",
-    "@angular/common": "18.2.13",
-    "@angular/compiler": "18.2.13",
-    "@angular/core": "18.2.13",
-    "@angular/forms": "18.2.13",
-    "@angular/platform-browser": "18.2.13",
-    "@angular/platform-browser-dynamic": "18.2.13",
-    "@angular/platform-server": "18.2.13",
-    "@angular/router": "18.2.13",
-    "@angular/ssr": "^18.2.12",
+    "@angular/animations": "19.0.5",
+    "@angular/common": "19.0.5",
+    "@angular/compiler": "19.0.5",
+    "@angular/core": "19.0.5",
+    "@angular/forms": "19.0.5",
+    "@angular/platform-browser": "19.0.5",
+    "@angular/platform-browser-dynamic": "19.0.5",
+    "@angular/platform-server": "19.0.5",
+    "@angular/router": "19.0.5",
+    "@angular/ssr": "^19.0.6",
     "angular-plotly.js": "^5.2.2",
     "file-saver": "^2.0.5",
     "happy-dom": "15.11.7",
     "plotly.js-dist-min": "^2.33.0",
     "rxjs": "^7.8.1",
-    "zone.js": "^0.14.3",
+    "zone.js": "^0.15.0",
     "@reflex/shared-types": "file:../../packages/reflex-shared-types",
     "express": "^4.18.2",
     "tslib": "^2.3.0",
     "@reflex/angular-components": "file:../../packages/reflex-angular-components"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^18.2.12",
-    "@angular/cli": "18.2.12",
-    "@angular/compiler-cli": "18.2.13",
-    "@angular/language-service": "18.2.13",
+    "@angular-devkit/build-angular": "^19.0.6",
+    "@angular/cli": "19.0.6",
+    "@angular/compiler-cli": "19.0.5",
+    "@angular/language-service": "19.0.5",
     "@gtvmbh/eslint-config": "^1.0.8",
     "@types/file-saver": "^2.0.7",
     "@types/jasmine": "~5.1.0",
@@ -62,10 +62,10 @@
     "karma-junit-reporter": "^2.0.1",
     "karma-spec-reporter": "^0.0.36",
     "karma-viewport": "^1.0.9",
-    "ng-packagr": "^18.2.1",
+    "ng-packagr": "^19.0.1",
     "puppeteer": "^23.11.1",
     "tslib": "^2.3.0",
-    "typescript": "5.4.5",
+    "typescript": "5.6.3",
     "@types/express": "^4.17.17"
   }
 }
