@@ -10,15 +10,14 @@ import { GestureTrackFrame } from '../data/gesture-track-frame';
 PlotlyModule.plotlyjs = PlotlyJS;
 
 @Component({
-  selector: 'app-timeline',
-  standalone: true,
-  imports: [PlotlyModule, 
-// TODO: `HttpClientModule` should not be imported into a component directly.
-// Please refactor the code to add `provideHttpClient()` call to the provider list in the
-// application bootstrap logic and remove the `HttpClientModule` import from this component.
-HttpClientModule, CommonModule],
-  templateUrl: './timeline.component.html',
-  styleUrls: ['./timeline.component.scss']
+    selector: 'app-timeline',
+    imports: [PlotlyModule,
+        // TODO: `HttpClientModule` should not be imported into a component directly.
+        // Please refactor the code to add `provideHttpClient()` call to the provider list in the
+        // application bootstrap logic and remove the `HttpClientModule` import from this component.
+        HttpClientModule, CommonModule],
+    templateUrl: './timeline.component.html',
+    styleUrls: ['./timeline.component.scss']
 })
 export class TimelineComponent implements OnInit {
   public max_value_layer = 1;
