@@ -3,11 +3,10 @@ import { CommonModule } from '@angular/common';
 import { TrackComponentComponent } from '../track-component/track-component.component';
 import { TimelineComponent } from '../timeline/timeline.component';
 @Component({
-  selector: 'app-pullup',
-  standalone: true,
-  imports: [CommonModule,TrackComponentComponent, TimelineComponent],
-  templateUrl: './pullup.component.html',
-  styleUrl: './pullup.component.scss'
+    selector: 'app-pullup',
+    imports: [CommonModule, TrackComponentComponent, TimelineComponent],
+    templateUrl: './pullup.component.html',
+    styleUrl: './pullup.component.scss'
 })
 export class PullupComponent {
   isPullupOpen: boolean = true;
